proto/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
markdown-it-py==3.0.0
mdurl==0.1.2
protobuf==5.29.2
Pygments==2.18.0
rich==13.9.4
typing_extensions==4.12.2
